Rearrange dependencies
