+++ interface/src/config.rs
@@ -7,10 +7,7 @@
     note = "Please use `crate::state::{DEFAULT_SLASH_PENALTY, DEFAULT_WARMUP_COOLDOWN_RATE}` instead"
 )]
 pub use super::state::{DEFAULT_SLASH_PENALTY, DEFAULT_WARMUP_COOLDOWN_RATE};
-use {
-    serde_derive::{Deserialize, Serialize},
-    solana_pubkey::declare_deprecated_id,
-};
+use solana_pubkey::declare_deprecated_id;
 
 // stake config ID
 declare_deprecated_id!("StakeConfig11111111111111111111111111111111");
@@ -19,7 +16,11 @@ declare_deprecated_id!("StakeConfig11111111111111111111111111111111");
     since = "1.16.7",
     note = "Please use `crate::state::warmup_cooldown_rate()` instead"
 )]
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, Copy)]
+#[cfg_attr(
+    feature = "serde",
+    derive(serde_derive::Deserialize, serde_derive::Serialize)
+)]
+#[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Config {
     /// how much stake we can activate/deactivate per-epoch as a fraction of currently effective stake
     pub warmup_cooldown_rate: f64,
